backend/pyproject.toml
[tool.poetry]
name = "oma-kulu-backend"
version = "0.1.0"
description = "Backend for Oma Kulu expense tracking application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
supabase = "^2.3.0"
pydantic = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 